mikrotik: modify the rb-750/4xx upgrade process to use procd's upgraded binary
This is because the routerboard has a watchdog that is triggered by
procd and when we kill procd the watchdog kicks in after a while, so
the scripted "upgraded" cannot work. procd's "upgraded" knows about
the watchdog and takes care to prevent it from rebooting the device,
so we should really use it.

This is done in a hacky way, introducing a new procd-tarball-upgrade
package that includes /sbin/upgraded but not /lib/upgrade/nand.sh that is
included in procd-nand. It would be nice if the solution could be unified,
but for the moment we only care about rb-750/4xx devices here, so procd-nand
can be left broken.