lib/netifd/proto/ppp.sh: disable pptp interface dependencies
In HWMN we use babel to distribute routes. Unfortunately, netifd
does not track the routing table of the kernel, but uses its own
routing table instead, which results in two problematic cases
regarding pptp:

1) There is no default gateway set up, so netifd thinks that the
pptp connection cannot work so it doesn't even try to bring the
interface up, giving no feedback as well.

2) There is a default gateway set up, in which case netifd will
start the pptp connection but it will also try to add a static
route to the pptp server through the default gateway, which is
terribly wrong since the route is different (distributed by babel).
And moreover, babeld will pick up the static route and distribute
it to the rest of the network, which will cause other connections
to that pptp server fail as well and packets will just loop in
the network.